requirements.txt
langchain==0.0.253
openai==0.27.8
python-dotenv==1.0.0
streamlit==1.24.0
faiss-cpu
tiktoken
openai